refactor: Update Value struct to use double for float values

+++ src/execution/executor_abstract.h
@@ -63,7 +63,7 @@ class AbstractExecutor {
                 break;
             }
             case TYPE_FLOAT: {
-                float fa = *(float *) a;
+                double fa = *(double *) a;
                 res.set_float(fa);
                 break;
             }
