+++ src/parser/ast.h
@@ -136,9 +136,9 @@ struct IntLit : public Value {
 };
 
 struct FloatLit : public Value {
-    float val;
+    double val;
 
-    FloatLit(float val_) : val(val_) {}
+    FloatLit(double val_) : val(val_) {}
 };
 
 struct StringLit : public Value {
@@ -258,7 +258,7 @@ struct SetStmt : public TreeNode {
 // Semantic value
 struct SemValue {
     int sv_int;
-    float sv_float;
+    double sv_float;
     std::string sv_str;
     bool sv_bool;
     OrderByDir sv_orderby_dir;
